requirements.txt
marshmallow==3.11.1
pyyaml==5.4.1
sanic==21.3.2
